Remove Begrunnelse from Tilgang DTO (#108)
Remove Begrunnelse from DTO object exposed to external consumers. It is removed to minimize the amount of information exposed externally about a Person and why a NavIdent is not granted access to the person, Begrunnelse has been removed from code in all consuming clients and can therefore be safely be removed.